requirements.txt
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.2
DateTime==4.4
future==0.18.2
netmiko==4.0.0
ntc-templates==3.0.0
paramiko==2.10.3
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
scp==0.14.4
six==1.16.0
tenacity==8.0.1
textfsm==1.1.2
zope.interface==5.4.0
